# apps/web-deprecated/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Vox Prismatic",
        "version": "1.0.0",
        "description": "Content Creation Platform API"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Local Development"
        }
    ],
    "security": [
        {
            "sanctum": []
        }
    ],
    "paths": {
...
        "/linkedin/callback": {
            "get": {
                "operationId": "linkedIn.callback_0",
                "tags": [
                    "LinkedIn"
                ],
                "parameters": [
                    {
                        "name": "code",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "state",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "enum": [
                                                "LinkedIn callback failed"
                                            ]
                                        },
                                        "code": {
                                            "type": "string",
                                            "enum": [
                                                "INVALID_INPUT"
                                            ]
                                        },
                                        "status": {
                                            "type": "integer",
                                            "enum": [
                                                400
                                            ]
                                        }
                                    },
                                    "required": [
                                        "error",
                                        "code",
                                        "status"
                                    ]
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "type": "string"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "connected": {
                                                    "type": "boolean"
                                                }
                                            },
                                            "required": [
                                                "connected"
                                            ]
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/linkedin/disconnect": {
            "post": {
                "operationId": "linkedIn.disconnect",
                "tags": [
                    "LinkedIn"
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "connected": {
                                            "type": "boolean"
                                        }
                                    },
                                    "required": [
                                        "connected"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/auth/linkedin/callback": {
            "get": {
                "operationId": "linkedIn.callback_0",
                "tags": [
                    "LinkedIn"
                ],
                "responses": {
                    "400": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "enum": [
                                                "LinkedIn callback failed"
                                            ]
                                        },
                                        "code": {
                                            "type": "string",
                                            "enum": [
                                                "INVALID_INPUT"
                                            ]
                                        },
                                        "status": {
                                            "type": "integer",
                                            "enum": [
                                                400
                                            ]
                                        }
                                    },
                                    "required": [
                                        "error",
                                        "code",
                                        "status"
                                    ]
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "type": "string"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "connected": {
                                                    "type": "boolean"
                                                }
                                            },
                                            "required": [
                                                "connected"
                                            ]
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/settings/profile": {
            "get": {
                "operationId": "settings.profile",
                "tags": [
                    "Settings"
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "user": {
                                            "type": "object",
                                            "properties": {
                                                "id": {
                                                    "type": "string"
                                                },
                                                "email": {
                                                    "type": "string"
                                                },
                                                "name": {
                                                    "type": "string"
                                                },
                                                "createdAt": {
                                                    "type": "string"
                                                },
                                                "isAdmin": {
                                                    "type": "boolean"
                                                },
                                                "stripeCustomerId": {
                                                    "type": "string"
                                                },
                                                "stripeSubscriptionId": {
                                                    "type": "string"
                                                },
                                                "subscriptionStatus": {
                                                    "type": "string"
                                                },
                                                "subscriptionPlan": {
                                                    "type": "string"
                                                },
                                                "subscriptionCurrentPeriodEnd": {
                                                    "type": "string"
                                                },
                                                "cancelAtPeriodEnd": {
                                                    "type": "boolean"
                                                },
                                                "trialEndsAt": {
                                                    "type": "string"
                                                },
                                                "trialNotes": {
                                                    "type": "string"
                                                }
                                            },
                                            "required": [
                                                "id",
                                                "email",
                                                "name",
                                                "createdAt",
                                                "isAdmin",
                                                "stripeCustomerId",
                                                "stripeSubscriptionId",
                                                "subscriptionStatus",
                                                "subscriptionPlan",
                                                "subscriptionCurrentPeriodEnd",
                                                "cancelAtPeriodEnd",
                                                "trialEndsAt",
                                                "trialNotes"
                                            ]
                                        }
                                    },
                                    "required": [
                                        "user"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            },
            "patch": {
                "operationId": "settings.updateProfile",
                "tags": [
                    "Settings"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": [
                                            "string",
                                            "null"
                                        ],
                                        "maxLength": 255
                                    },
                                    "email": {
                                        "type": [
                                            "string",
                                            "null"
                                        ],
                                        "format": "email"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "user": {
                                            "type": "object",
                                            "properties": {
                                                "id": {
                                                    "type": "string"
                                                },
                                                "email": {
                                                    "type": "string"
                                                },
                                                "name": {
                                                    "type": "string"
                                                },
                                                "createdAt": {
                                                    "type": "string"
                                                },
                                                "isAdmin": {
                                                    "type": "boolean"
                                                },
                                                "stripeCustomerId": {
                                                    "type": "string"
                                                },
                                                "stripeSubscriptionId": {
                                                    "type": "string"
                                                },
                                                "subscriptionStatus": {
                                                    "type": "string"
                                                },
                                                "subscriptionPlan": {
                                                    "type": "string"
                                                },
                                                "subscriptionCurrentPeriodEnd": {
                                                    "type": "string"
                                                },
                                                "cancelAtPeriodEnd": {
                                                    "type": "boolean"
                                                },
                                                "trialEndsAt": {
                                                    "type": "string"
                                                },
                                                "trialNotes": {
                                                    "type": "string"
                                                }
                                            },
                                            "required": [
                                                "id",
                                                "email",
                                                "name",
                                                "createdAt",
                                                "isAdmin",
                                                "stripeCustomerId",
                                                "stripeSubscriptionId",
                                                "subscriptionStatus",
                                                "subscriptionPlan",
                                                "subscriptionCurrentPeriodEnd",
                                                "cancelAtPeriodEnd",
                                                "trialEndsAt",
                                                "trialNotes"
                                            ]
                                        }
                                    },
                                    "required": [
                                        "user"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Cookie-based session authentication using Laravel Sanctum via content_session cookie. For non-GET requests, fetch /sanctum/csrf-cookie and include X-XSRF-TOKEN header.",
                "in": "cookie",
                "name": "content_session"
            }
        },
...
        "responses": {
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            }
        }
    }
}